Clean up backend specific types
