+++ server/src/types/apiError.ts
