+++ server/src/types/general.ts
@@ -2,15 +2,22 @@
 //
 // SPDX-License-Identifier: MIT
 
-import {SystemRole} from '@common/types';
 import {z} from 'zod';
+
+import {CourseRoleType, SystemRole} from '@common/types';
+import Course from '../database/models/course';
+import CourseTranslation from '../database/models/courseTranslation';
+import User from '../database/models/user';
 
 export const stringToIdSchema = z
   .string()
   .regex(/^\d+$/)
   .pipe(z.coerce.number().int().min(1));
 
-export interface JwtClaims {
-  role: SystemRole;
-  id: number;
-}
+export type JwtClaims = {role: SystemRole; id: number};
+
+type UserWithRole = User & {CourseRole: {role: CourseRoleType}};
+export type CourseFull = Course & {
+  CourseTranslations: CourseTranslation[];
+  Users: UserWithRole[];
+};
